refactor: reducer.cpp をドメイン別ファイルに分割 (Phase 3)
reducer.cpp (1094行→294行) を scroll / search / input / navigation / ui / system の
6ドメインに分割。std::visit ディスパッチと横断共有ヘルパーは reducer.cpp に残し、
宣言は reducer_internal.h に集約。

ロジック・AppAction の variant 構造は不変（コード移動と宣言整備のみ）
